+++ changelog.md
@@ -2,6 +2,8 @@
 
 ## Version 4
 
+### Version 4.0.12
+
 ### Version 4.0.11
 
 * `quasar-utils`: Add quasar-core to cordaRuntime. This ensures that it is excluded from "fat jar" CorDapps [CORDA-1301].